requirements.txt
streamlit
pillow
numpy
easyocr
torch
torchvision
scikit-image
imageio
gspread
google-auth
google-auth-oauthlib
google-api-python-client
requests
